adaptlp/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
python-dotenv==1.0.0
google-generativeai==0.7.2
beautifulsoup4==4.12.3
lxml==5.2.2
pillow==10.4.0
pydantic==2.8.0
python-multipart==0.0.9
aiofiles==24.1.0
pytest==7.4.0
pytest-asyncio==0.21.1
